875 pi small types

+++ KioChess/Engine/Models/Config/EndGameConfiguration.cs
@@ -4,8 +4,8 @@ namespace Engine.Models.Config
 {
     public class EndGameConfiguration : IEndGameConfiguration
     {
-        public int MaxEndGameDepth { get; set; }
+        public short MaxEndGameDepth { get; set; }
 
-        public int EndGameDepthOffset { get; set; }
+        public short EndGameDepthOffset { get; set; }
     }
 }
