+++ KioChess/Engine/Strategies/Models/IterativeDeepingModel.cs
@@ -4,7 +4,7 @@ namespace Engine.Strategies.Models
 {
     public class IterativeDeepingModel
     {
-        public int Depth { get; set; }
+        public sbyte Depth { get; set; }
         public StrategyBase Strategy { get; set; }
     }
 }
